==> 4a-logistic_regression.ipynb <==
import pandas as pd
df = pd.read_csv("review_history_with_time_features.csv")
df

from sklearn.metrics import roc_auc_score
roc_auc_score(df['was_remembered'], df['time_since_prev_review_standard'])

from sklearn.model_selection import train_test_split

def split(df, test_size = 0.2):
  unique_card_ids = df['card_id'].unique()

  train_card_ids, test_card_ids = train_test_split(unique_card_ids, test_size=test_size, random_state=42)
  df_train = df[df['card_id'].isin(train_card_ids)]
  df_test = df[df['card_id'].isin(test_card_ids)]
  return df_train, df_test

df_train, df_test = split(df, .15)

from sklearn.linear_model import LogisticRegression
lr = LogisticRegression()
lr.fit(df_train[['time_since_prev_review_standard']], df_train['was_remembered'])

preds = lr.predict_proba(df_test[['time_since_prev_review_standard']])

roc_auc_score(df_test["was_remembered"], preds[:, 1])
